btsh.md
Bitsun is a financial network based on the chain reaction agreement, which aims to promote and develop the fastest and cheapest payment and settlement system worldwide. It allows payments in various currencies as easy as sending an email.
As the core component of Bitsunny Network, the underlying Bitsunny is a series of servers running point-to-point software. Anyone is free to set up a BitShengyou server to accept or pay, and the reward plan is sent or redeemed every day to assign reservations to define each qualified user. The use of BitSunny is free. The BitSunny network is owned by every participant.
This is the holy grail of Besyshlad, citing a repository of P2P servers.
Repository contents for BitSheng You integrator n. / SBI nScigh and data files.
./Generate intermediate and final build output.

./ Generate nPlatform or IDE-specific project files.

./kageDoubnnmet and a sample configuration file.

. / scr n Source code directory. Some of the directories included here are external repositories inlined through git-subtrees, see the corresponding readme for details.

Bit Sanyo